requirements.txt
torch>=2.1.0+cu121
torch>=2.0.0
torchvision>=0.16.0+cu121
torchaudio>=2.1.0+cu121

# Hugging Face ecosystem
transformers>=4.30.0
datasets>=2.14.0
 accelerate>=0.20.0

# Model utilities
peft>=0.5.0          # Parameter-Efficient Fine-Tuning
bitsandbytes>=0.41.0 # Low-bit quantization

# Computer vision
Pillow>=10.0.0
opencv-python>=4.8.0

# Training and evaluation
tqdm>=4.65.0
tensorboard>=2.13.0

# Notebook visualization
matplotlib>=3.7.0
seaborn>=0.12.0

# Optional: for advanced use cases
open_clip_torch>=2.23.0
sentencepiece>=0.1.99
